lesson-4/dedicated-algorithm/requirements.txt
Faker==37.12.0
greenlet==3.2.4
psycopg2-binary==2.9.11
python-dotenv==1.2.1
typing_extensions==4.15.0
tzdata==2025.2
